fix: compare onClick references instead of 'toString()'
Previously, I was comparing the onClick `toString()`. This creates a bug
because, if the onClick function uses variables in its scope, their new
values are not available when running the onClick() (thus, always the
initial value will be used).
(`() => {console.log(a)}` is seen as identical, even if the value of `a`
changes).
The solution is to compare references instead.
As a consequence, the event listener is udpated when the scope of
onClick changes.
NB: Another solution could have been to keep with `toString()` and add
the scoped variables in the comparison. However, I couldn't find a way
to do it).